+++ NEWS.md
@@ -1,6 +1,9 @@
 shinytest 1.4.0.9002
 ===============
 
+* New `Widget$getHtml()` returns the complete HTML of the selected widget 
+  (#347).
+
 * `ShinyDriver$intialize()` gains a `renderArgs` method which is passed along
   to `rmarkdown::run()` for interactive Rmds (#249); this makes it possible
   to set parameters for parameterised reports.